Challenges/TamuDatathon/TamuDatathon SourceCode/src/requirements.txt
numpy
keras
tensorflow==2.10.0
pillow==9.2.0
pandas